+++ src/app/streckengrafik/services/sg-1-load-trainrun-item.service.ts
@@ -172,37 +172,10 @@ export class Sg1LoadTrainrunItemService implements OnDestroy {
       []
     );
 
-    // create a new graph object
-    const graph = new MultiSelectNodeGraph(this.nodeService);
-    const edgeList = [];
-
-    // retrieve edges
-    nodes.forEach((node1) => {
-      nodes.forEach((node2) => {
-        const n1 = node1.getId() < node2.getId() ? node1.getId() : node2.getId();
-        const n2 = node1.getId() < node2.getId() ? node2.getId() : node1.getId();
-        const ts12 = this.trainrunSectionService.getTrainrunSections().find((ts: TrainrunSection) =>
-          (ts.getSourceNodeId() === n1 && ts.getTargetNodeId() === n2)
-        );
-        if (ts12 !== undefined) {
-          if (edgeList.find((a) => (a[0] === n1 && a[1] === n2)) === undefined) {
-            edgeList.push([n1, n2]);
-          }
-        } else {
-          const ts21 = this.trainrunSectionService.getTrainrunSections().find((ts: TrainrunSection) =>
-            (ts.getSourceNodeId() === n2 && ts.getTargetNodeId() === n1)
-          );
-          if (ts21 !== undefined) {
-            if (edgeList.find((a) => (a[0] === n2 && a[1] === n1)) === undefined) {
-              edgeList.push([n2, n1]);
-            }
-          }
-        }
-      });
-    });
-
-    // insert all edges (graph)
-    graph.createAdjList(edgeList);
+    // create a new graph object from nodes
+    const graph =
+      new MultiSelectNodeGraph(this.nodeService, this.trainrunSectionService);
+    graph.convertNetzgrafikSubNodesToGraph(nodes);
 
     // get all starting ending nodes (degree == 1)
     const singlePath = graph.getStartEndingVertices();
@@ -251,12 +224,14 @@ export class Sg1LoadTrainrunItemService implements OnDestroy {
       });
     }
 
+    // convert corridor to path elements
     corridor.forEach((n, idx, nodArray) => {
       if (idx > 0) {
         this.makePathElement(nodArray[idx - 1].getId(), nodArray[idx].getId());
       }
     });
 
+    // clean up
     this.trainrunService.unselectAllTrainruns(false);
     this.nodeService.unselectAllNodes(false);
   }
